+++ src/lib/settings/SchemaPiece.js
@@ -43,7 +43,6 @@ class SchemaPiece extends Schema {
 		 * The type of this key.
 		 * @since 0.5.0
 		 * @type {string}
-		 * @name SchemaPiece#type
 		 */
 		this.type = options.type.toLowerCase();
 
@@ -51,7 +50,6 @@ class SchemaPiece extends Schema {
 		 * Whether this key should store multiple or a single value.
 		 * @since 0.5.0
 		 * @type {boolean}
-		 * @name SchemaPiece#array
 		 */
 		this.array = 'array' in options ? options.array : Array.isArray(options.default);
 
@@ -59,7 +57,6 @@ class SchemaPiece extends Schema {
 		 * What this key should provide by default.
 		 * @since 0.5.0
 		 * @type {*}
-		 * @name SchemaPiece#default
 		 */
 		this.default = 'default' in options ? options.default : this._generateDefault();
 
@@ -67,7 +64,6 @@ class SchemaPiece extends Schema {
 		 * The minimum value for this key.
 		 * @since 0.5.0
 		 * @type {?number}
-		 * @name SchemaPiece#min
 		 */
 		this.min = 'min' in options ? options.min : null;
 
@@ -75,23 +71,20 @@ class SchemaPiece extends Schema {
 		 * The maximum value for this key.
 		 * @since 0.5.0
 		 * @type {?number}
-		 * @name SchemaPiece#max
 		 */
 		this.max = 'max' in options ? options.max : null;
 
 		/**
 		 * A tuple of strings containing the path and the datatype.
 		 * @since 0.5.0
-		 * @type {string[]}
-		 * @name SchemaPiece#sql
+		 * @type {string}
 		 */
-		this.sql = [this.path, null];
+		this.sql = null;
 
 		/**
 		 * Whether this key should be configurable by the config command. When type is any, this key defaults to false.
 		 * @since 0.5.0
 		 * @type {boolean}
-		 * @name SchemaPiece#configurable
 		 */
 		this.configurable = 'configurable' in options ? options.configurable : this.type !== 'any';
 
@@ -143,15 +136,15 @@ class SchemaPiece extends Schema {
 		if (!isObject(options)) throw new TypeError(`SchemaPiece#edit expected an object as a parameter. Got: ${typeof options}`);
 
 		const edited = new Set();
-		if (typeof options.sql === 'string' && this.sql[1] !== options.sql) {
-			this.sql[1] = options.sql;
+		if (typeof options.sql === 'string' && this.sql !== options.sql) {
+			this.sql = options.sql;
 			edited.add('SQL');
 			if (this.gateway.sql) await this.gateway.provider.updateColumn(this.gateway.type, this.path, options.sql);
 		}
 		if (typeof options.default !== 'undefined' && this.default !== options.default) {
 			this._schemaCheckDefault({ ...this.toJSON(), ...options });
 			this.default = options.default;
-			if (!edited.has('SQL')) this.sql[1] = this._generateSQLDatatype(options.sql);
+			if (!edited.has('SQL')) this.sql = this._generateSQLDatatype(options.sql);
 			edited.add('DEFAULT');
 		}
 		if (typeof options.min !== 'undefined' && this.min !== options.min) {
@@ -272,17 +265,21 @@ class SchemaPiece extends Schema {
 		if (!provider.sql || !provider.qb) return null;
 
 		const qb = provider.qb.create();
-		switch (this.type) {
-			case 'integer': qb.setType(['INTEGER', 'INT'], this.max).setNotNull();
-				break;
-			case 'float': qb.setType(['FLOAT', 'REAL'], this.max).setNotNull();
-				break;
-			case 'boolean': qb.setType('BOOLEAN').setNotNull();
-				break;
-			case 'string': qb.setType(['VARCHAR', 'TEXT'], this.max);
-				break;
-			default: qb.setType('TEXT', this.max);
-				break;
+		if (this.array) {
+			qb.setType(['JSON', 'TEXT']);
+		} else {
+			switch (this.type) {
+				case 'integer': qb.setType(['INTEGER', 'INT'], this.max).setNotNull();
+					break;
+				case 'float': qb.setType(['FLOAT', 'REAL'], this.max).setNotNull();
+					break;
+				case 'boolean': qb.setType('BOOLEAN').setNotNull();
+					break;
+				case 'string': qb.setType(['VARCHAR', 'TEXT'], this.max);
+					break;
+				default: qb.setType('TEXT', this.max);
+					break;
+			}
 		}
 		return qb.setDefault(SchemaPiece._parseSQLValue(this.default))
 			.toString();
@@ -299,7 +296,7 @@ class SchemaPiece extends Schema {
 		if (typeof object.default !== 'undefined') this.default = object.default;
 		if (typeof object.min === 'number') this.min = object.min;
 		if (typeof object.max === 'number') this.max = object.max;
-		if (typeof object.sql === 'string') this.sql[1] = object.sql;
+		if (typeof object.sql === 'string') this.sql = object.sql;
 		if (typeof object.configurable === 'boolean') this.configurable = object.configurable;
 	}
 
@@ -317,7 +314,7 @@ class SchemaPiece extends Schema {
 
 		// Check if the 'options' parameter is an object.
 		if (!isObject(options)) throw new TypeError(`SchemaPiece#init expected an object as a parameter. Got: ${typeof options}`);
-		this.sql[1] = this._generateSQLDatatype(options.sql);
+		this.sql = this._generateSQLDatatype(options.sql);
 
 		return true;
 	}
@@ -334,7 +331,7 @@ class SchemaPiece extends Schema {
 			default: this.default,
 			min: this.min,
 			max: this.max,
-			sql: this.sql[1],
+			sql: this.sql,
 			configurable: this.configurable
 		};
 	}